+++ test/unit/mysql_db_admin/main.py
@@ -45,6 +45,7 @@ class ArgParser(object):
         get_val
         arg_xor_dict
         insert_arg
+        arg_parse2
 
     """
 
@@ -73,6 +74,7 @@ class ArgParser(object):
         self.opt_or2 = True
         self.opt_xor_val = None
         self.opt_xor_val2 = True
+        self.argparse2 = True
 
     def arg_cond_req(self, opt_con_req):
 
@@ -193,6 +195,18 @@ class ArgParser(object):
 
         self.args_array[arg_key] = arg_val
 
+    def arg_parse2(self):
+
+        """Method:  arg_parse2
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_parse2.
+
+        Arguments:
+
+        """
+
+        return self.argparse2
+
 
 class ProgramLock(object):
 
@@ -229,6 +243,8 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
+        test_arg_parse2_false
+        test_arg_parse2_true
         test_help_true
         test_help_false
         test_arg_req_false
@@ -260,6 +276,40 @@ class UnitTest(unittest.TestCase):
 
         self.proglock = ProgramLock(["cmdline"], "FlavorID")
         self.args = ArgParser()
+
+    @mock.patch("mysql_db_admin.gen_class.ArgParser")
+    def test_arg_parse2_false(self, mock_arg):
+
+        """Function:  test_arg_parse2_false
+
+        Description:  Test arg_parse2 returns false.
+
+        Arguments:
+
+        """
+
+        self.args.argparse2 = False
+
+        mock_arg.return_value = self.args
+
+        self.assertFalse(mysql_db_admin.main())
+
+    @mock.patch("mysql_db_admin.gen_libs.help_func")
+    @mock.patch("mysql_db_admin.gen_class.ArgParser")
+    def test_arg_parse2_true(self, mock_arg, mock_help):
+
+        """Function:  test_arg_parse2_true
+
+        Description:  Test arg_parse2 returns true.
+
+        Arguments:
+
+        """
+
+        mock_arg.return_value = self.args
+        mock_help.return_value = True
+
+        self.assertFalse(mysql_db_admin.main())
 
     @mock.patch("mysql_db_admin.gen_libs.help_func")
     @mock.patch("mysql_db_admin.gen_class.ArgParser")